Fix time slice connection checks applying when interconnecting slices.

+++ Assets/Scripts/Backend/Simulation/World/BackendImpl.cs
@@ -50,33 +50,45 @@ namespace Backend.Simulation.World
             var sliceA = getTimeSliceOfNodeByGuid(a);
             var sliceB = getTimeSliceOfNodeByGuid(b);
             //TODO connecting generator and node in different slices only works node -> generator not generator -> node
-            if (sliceA == null || sliceB == null)
-                return null;
-
-            var grid = sliceA.TimeSliceGrid;
-            int numOfConnectionsCrossed = 0;
-            foreach (var cell in cellsOfConnection)
+            if (sliceA == null)
             {
-                if (grid.IsCellOccupied(cell, out var node, out var connection))
-                {
-                    if (node != null && (!node.guid.Equals(a) && !node.guid.Equals(b)))
-                    {
-                        Debug.Log("Cannot link because there is a node in its path in slices: "+sliceA.SliceNumber+"-"+sliceB.SliceNumber);
-                        return null;
-                    }
+                Debug.Log("Error while linking nodes -> Slice A could not be found");
+                return null;
+            }
+            if (sliceB == null)
+            {
+                Debug.Log("Error while linking nodes -> Slice B could not be found");
+                return null;
+            }
 
-                    if (connection != null)
+            // Only check for occupied cells if we stay in one slice. Otherwise, this should not be used.
+            if (sliceA.Equals(sliceB))
+            {
+                var grid = sliceA.TimeSliceGrid;
+                int numOfConnectionsCrossed = 0;
+                foreach (var cell in cellsOfConnection)
+                {
+                    if (grid.IsCellOccupied(cell, out var node, out var connection))
                     {
-                        numOfConnectionsCrossed++;
-                        if (numOfConnectionsCrossed > bridgesBuilt)
+                        if (node != null && (!node.guid.Equals(a) && !node.guid.Equals(b)))
                         {
-                            Debug.Log("Cannot link because there are not enough bridges built to cross connections in slices: "+sliceA.SliceNumber+"-"+sliceB.SliceNumber);
+                            Debug.Log("Cannot link because there is a node in its path in slices: "+sliceA.SliceNumber+"-"+sliceB.SliceNumber);
                             return null;
+                        }
+
+                        if (connection != null)
+                        {
+                            numOfConnectionsCrossed++;
+                            if (numOfConnectionsCrossed > bridgesBuilt)
+                            {
+                                Debug.Log("Cannot link because there are not enough bridges built to cross connections in slices: "+sliceA.SliceNumber+"-"+sliceB.SliceNumber);
+                                return null;
+                            }
                         }
                     }
                 }
             }
-
+            
             var connectionId = _storage.link(a, b, cellsOfConnection);
             if (connectionId == null)
             {
@@ -87,19 +99,24 @@ namespace Backend.Simulation.World
 
             var connectionObj = _storage.guidToConnections[(Guid)connectionId];
 
-            bool reserved = grid.TryAddConnectionCells(
-                connectionObj,
-                cellsOfConnection,
-                _storage.guidToNodesMapping[a],
-                _storage.guidToNodesMapping[b],
-                bridgesBuilt
+            if (sliceA.Equals(sliceB))
+            {
+                var grid = sliceA.TimeSliceGrid;
+                bool reserved = grid.TryAddConnectionCells(
+                    connectionObj,
+                    cellsOfConnection,
+                    _storage.guidToNodesMapping[a],
+                    _storage.guidToNodesMapping[b],
+                    bridgesBuilt
                 );
 
-            if (!reserved)
-            {
-                _storage.unlink((Guid)connectionId);
-                return null;
+                if (!reserved)
+                {
+                    _storage.unlink((Guid)connectionId);
+                    return null;
+                }
             }
+            
             _storage.recalculatePaths();
             return connectionObj.guid;
         }
